[golinks] introduce '+' handling
In typical shorturl installations it possible to append a '+' to the URL
and see what the URL is along with any metadata associated with it.

While the current routing mechanism makes this difficult, begin the
introduction

This was tested manually as I could not determine any existing test
infrastructure for the web framework.